+++ src/main/java/com/prodemy/pembayaran/listrik/model/entity/JenisPelanggan.java
@@ -17,7 +17,7 @@ public class JenisPelanggan {
     private String daya;
 
     @Column
-    private Long tarif;
+    private Double tarif;
 
     public Long getIdJenis() {
         return idJenis;
@@ -43,11 +43,11 @@ public class JenisPelanggan {
         this.daya = daya;
     }
 
-    public Long getTarif() {
+    public Double getTarif() {
         return tarif;
     }
 
-    public void setTarif(Long tarif) {
+    public void setTarif(Double tarif) {
         this.tarif = tarif;
     }
 }
